Separated Maps from levels.h
Separated the game maps from levels.h. Next step: Make a new test map to test loading between levels.
