+++ gameFiles/levels/levels.h
@@ -2,6 +2,7 @@
 #define dos_obj_levels
 
 #include "level.h"
+#include "map.h"
 
 // This is where DoS would put its generated levels.
 #include "levelExample.h"
@@ -22,33 +23,6 @@
 #define LEV_MAIN_INTROLEVEL 5
 #define LEV_JUNG_RAINHALLWAY 6
 #define LEV_TEST_MULTILIGHTS 7
-
-struct LevLoaded{
-   Level* lev;
-   bool loaded;
-};
-
-class Map{
-   private:
-      double x;
-      double y;
-      double prevCX;
-      double prevCY;
-      double w;
-      double h;
-      bool firstLoad;
-      /// TODO: Find a better way to load levels besides iterating through every level.
-      std::vector<Level *> levels;
-      std::vector<LevLoaded> getLevelsInArea(double p, double q1, double q2, bool horizontal, std::vector<LevLoaded> prev);
-      int superMapID;
-   public:
-      void addLevel(Level* l, double X, double Y);
-      std::vector<Level *> updateLoadedLevels(LevelList* l, GLUtil* glu);
-      std::vector<Level *> updateLoadedLevels(LevelList* l, double X, double Y, double W, double H);
-      Map();
-      int getSuperMapID(){ return superMapID; };  
-      bool inBounds(double X, double Y, double W, double H);
-};
 
 /// TODO: Find a better home for this.
 template <class T>
